requirements.txt
beautifulsoup4
coverage
django
django-cors-headers
django-debug-toolbar
django-filter
django-redis
djangorestframework
flake8
flake8-isort
influxdb
pandas
piprot
psycopg2-binary
pytest
pytest-django
pyuwsgi
redis
requests
requests-cache
